--- motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/admin/persentation/dto/request/ApproveMusicRequest.kt
@@ -1,5 +1,0 @@
-package kr.motung_i.backend.domain.admin.persentation.dto.request
-
-data class ApproveMusicRequest(
-    val rankNumber: Int
-)
