+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/music/presentation/dto/request/ApproveMusicRequest.kt
@@ -1,0 +1,5 @@
+package kr.motung_i.backend.domain.music.presentation.dto.request
+
+data class ApproveMusicRequest(
+    val rankNumber: Int
+)
