+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/music/presentation/dto/response/impl/ApprovedMusicResponse.kt
@@ -1,5 +1,6 @@
-package kr.motung_i.backend.domain.music.presentation.dto.response
+package kr.motung_i.backend.domain.music.presentation.dto.response.impl
 
+import kr.motung_i.backend.domain.music.presentation.dto.response.MusicResponse
 import kr.motung_i.backend.persistence.music.entity.Music
 
 class ApprovedMusicResponse(
